+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 25 July 2024
 - More bugfixes for The War Within.
+- Added support for cooldowns on Paladins' Flash of Light spell.
 
 # 24 July 2024
 - Various bugfixes for The War Within.